标准化

+++ app/src/main/cpp/common/elf_op.h
@@ -62,7 +62,7 @@ private:
   void* dynstr; // 名称字符串表(Section)
 
   void* dynsym;       // 符号表(Section)
-  unsigned int nsyms; // 符号表中的符号item条数
+  size_t nsyms; // 符号表中的符号item条数
 
   off_t bias; // 是节头表在进程地址空间中的基地址(偏移地址) // TODO: why ?
